See ChangeLog for 2004-01-21. Lots of changes, including working stream execution. Hoorah\!
git-svn-id: svn://svn.code.sf.net/p/libsh/code/sh/trunk@345 afdca40c-03d6-0310-8ede-e9f093b21075

+++ src/ShConstant.hpp
@@ -52,10 +52,10 @@ private:
   ShConstant<N, T>& operator=(const ShConstant<N, T>& other);
 };
 
-typedef ShConstant<1, double> ShConstant1f;
-typedef ShConstant<2, double> ShConstant2f;
-typedef ShConstant<3, double> ShConstant3f;
-typedef ShConstant<4, double> ShConstant4f;
+typedef ShConstant<1, float> ShConstant1f;
+typedef ShConstant<2, float> ShConstant2f;
+typedef ShConstant<3, float> ShConstant3f;
+typedef ShConstant<4, float> ShConstant4f;
 
 }
